Fix Foodcritic issues
Summary:
Handles three issues with foodcritic:
- Fixes the problem referenced in T772 by setting default severity to warning.
- Adds the description to the Foodcritic linter message.
- Excludes FC011, FC031, and FC045 from the foodcritic linting.

The above three rules unfortunately throw false positives with the way FC is
used by Arcanist, and trigger for files that do not exist (and which should not
exist) in the parent directory, templates directory, etc. This is particularly
a problem with 'arc lint --everything'. Excluding the rules seems like a better
solution than having arc lint broken.

A better solution that checks that the missing files are the correct ones (i.e.,
located in the root cookbook directory) could possibly be implemented later.

Test Plan:
Run `arc lint --everything` on any Chef cookbook with arcanist-support patched
up to this version.

Reviewers: #phabricator, kimtj

Reviewed By: #phabricator, kimtj

Subscribers: kimtj

Maniphest Tasks: T772

Differential Revision: https://phab.met.no/D568