SBTUnitTest corrected for standard Play projects
Summary:
Existing version of the SBTTestEngine looks for test reports in a non-standard directory.
This caused an exception to be thrown. Changes fix that issue.

Also fixed some lint issues (and switched off a Pear style that we do not want to use).

Test Plan: Check out rMETAPIAUTH and run `arc unit` in the repository. Verify that there are no exceptions thrown.

Reviewers: #phabricator, borgem

Reviewed By: borgem

Subscribers: davidb, borgem

Differential Revision: https://phab.met.no/D579