Added a working cpplint linter
Summary:
The base arcanist implementation for cpplint currently doesn't work; see https://secure.phabricator.com/T8404.

Until the latter is in a working state, this will serve in the meantime.

Note: The __phutil__ files are generated by `arcanist liberate`; they do not need to be reviewed in detail.

Test Plan:
Install cpplint.py

Add the following to an .arclint file in a project.

```
    "googleccplint": {
      "type": "googlecpplint",
      "include": [
        "(\\.cpp$)",
        "(\\.cc$)",
        "(\\.c$)",
        "(\\.hpp$)",
        "(\\.h$)"
      ]
    }
```

'arc lint` on a C++ file

Reviewers: #phabricator, vegardb, borgem

Reviewed By: borgem

Differential Revision: https://phab.met.no/D1291