Improved arcanist support scripts
Summary:
Making arcanist easier to install and maintain for users. Debian package works poorly, and hard to maintain (lack servers and no one to work on it). Installing and maintaining directly from git is similar to what others do, and seems more in-tune with the resources we seem willing to spend on this.

- Added arc-install to simplify installation of arcanist and libphutil.
- Added arc-upgrade to simplify upgrading of arcanist and libphutil.
- Added arc-uninstall to... guess.
- Renamed arc_delete_patches to arc-clean to follow standard linux script naming conventions.

Note that the scripts won't work in the current state. since they are dependent on executables being present in git.

Test Plan:
Clone the arcanist-support repository and patch it with this diff.

From the diff,

- Run `bin/arc-install -i DIR1`. Where DIR1 is an existing directory. Verify that files are installed.

- Run `bin/arc-install -i DIR1` again.  Verify that it fails.

Copy executables from bin/ to DIR1/include/php/arcanist-support/bin

- Run `DIR1/bin/arc-upgrade`.  Verify that upgrade worked.

- Run `DIR1/bin/arc-clean` on a patched repository. Verify that patch branches are deleted.

- Run `DIR1/bin/arc-uninstall`.  Verify that all files are deleted.

Reviewers: #phabricator, alexanderb

Reviewed By: alexanderb

Differential Revision: https://phab.met.no/D1294